refactor(coin): controller 패키지를 presentation 패키지로 이름 변경
