+++ src/main/java/com/project/bumawiki/domain/coin/presentation/dto/PriceResponse.java
@@ -1,4 +1,4 @@
-package com.project.bumawiki.domain.coin.controller.dto;
+package com.project.bumawiki.domain.coin.presentation.dto;
 
 import java.time.LocalDateTime;
